[fix] pendingReservation - 간병 확정하기 코드 분리 및 디버깅
